requirements.txt
Eel
playsound==1.2.2
pyttsx3
openai-whisper
sounddevice
pandas
numpy==1.26.4
pywhatkit
pvporcupine==1.9.5
pyaudio
open-cv-python
opencv-contrib-python
openai

tensorflow==2.10.0
joblib

nltk
contractions
scikit-learn==1.5.2

librosa

python-dotenv